requirements.txt
langchain==0.0.251
streamlit==1.25.0
openai==0.27.4
faiss-cpu==1.7.4
tiktoken==0.3.3
xmltodict==0.13.0
bs4
pinecone-client
git+https://github.com/tomquirk/linkedin-api.git